% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_decode}
\alias{base64_decode}
\title{Decode data from the base64 encoding.}
\usage{
base64_decode(x)
}
\arguments{
\item{x}{A string. If \code{file.exist(x)} returns \code{TRUE}, it is assumed to be a file or a connection
to be read as string via \code{readChar()}.}
}
\value{
A raw vector.
}
\description{
Decode a file or a string from the base64 encoding.
}
\examples{
xfun::base64_decode("AQIDBAUGBwgJCg==")
}
