% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{base_pkgs}
\alias{base_pkgs}
\title{Get base R package names}
\usage{
base_pkgs()
}
\value{
A character vector of base R package names.
}
\description{
Return names of packages from \code{\link[=installed.packages]{installed.packages()}} of which the
priority is \code{"base"}.
}
\examples{\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
xfun::base_pkgs()
\dontshow{\}) # examplesIf}
}
