% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{bg_process}
\alias{bg_process}
\title{Start a background process}
\usage{
bg_process(command, args = character(), timeout = 30)
}
\arguments{
\item{command, args}{The system command and its arguments, to be passed to
\code{\link{system2}()}.}

\item{timeout}{The limit of the elapsed time (in seconds) to wait for the
process to start and find its process ID.}
}
\value{
The process ID as a character string.
}
\description{
Use \code{\link{system2}(wait = FALSE)} to start a background process, and
return its process ID.
}
\seealso{
\code{\link{proc_kill}()} to kill a process.
}
