% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{broken_packages}
\alias{broken_packages}
\title{Find out broken packages and reinstall them}
\usage{
broken_packages(reinstall = TRUE)
}
\arguments{
\item{reinstall}{Whether to reinstall the broken packages, or only list their
names.}
}
\value{
A character vector of names of broken package.
}
\description{
If a package is broken (i.e., not \code{\link[=loadable]{loadable()}}), reinstall it.
}
\details{
Installed R packages could be broken for several reasons. One common reason
is that you have upgraded R to a newer \code{x.y} version, e.g., from \verb{4.0.5} to
\verb{4.1.0}, in which case you need to reinstall previously installed packages.
}
