% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{bump_version}
\alias{bump_version}
\title{Bump version numbers}
\usage{
bump_version(x)
}
\arguments{
\item{x}{A vector of version numbers (of the class \code{"numeric_version"}),
or values that can be coerced to version numbers via
\code{as.numeric_version()}.}
}
\value{
A vector of new version numbers.
}
\description{
Increase the last digit of version numbers, e.g., from \code{0.1} to
\code{0.2}, or \code{7.23.9} to \code{7.23.10}.
}
\examples{
xfun::bump_version(c("0.1", "91.2.14"))
}
