% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{dir_create}
\alias{dir_create}
\title{Create a directory recursively by default}
\usage{
dir_create(x, recursive = TRUE, ...)
}
\arguments{
\item{x}{A path name.}

\item{recursive}{Whether to create all directory components in the path.}

\item{...}{Other arguments to be passed to \code{\link{dir.create}()}.}
}
\value{
A logical value indicating if the directory either exists or is
  successfully created.
}
\description{
First check if a directory exists. If it does, return \code{TRUE}, otherwise
create it with \code{\link{dir.create}(recursive = TRUE)} by default.
}
