% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{file_path}
\alias{file_path}
\title{Construct path to file}
\usage{
file_path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors used to construct the path.}

\item{fsep}{the path separator to use (assumed to be ASCII).}
}
\description{
A wrapper function of \code{file.path} which preserves a trailing slash on Windows.
}
\examples{
library(xfun)
file_path("mydir1", "mydir2/")
}
