% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{file_string}
\alias{file_string}
\title{Read a text file and concatenate the lines by \code{'\\n'}}
\usage{
file_string(file)
}
\arguments{
\item{file}{Path to a text file (should be encoded in UTF-8).}
}
\value{
A character string of text lines concatenated by \code{'\\n'}.
}
\description{
The source code of this function should be self-explanatory.
}
\examples{
xfun::file_string(system.file("DESCRIPTION", package = "xfun"))
}
