% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_paste_insert.R
\name{grab_clipboard}
\alias{grab_clipboard}
\title{grab image in clipboard (if have), and save to a local filepath}
\usage{
grab_clipboard(filepath)
}
\arguments{
\item{filepath}{a string, save clipboard's image in this filepath}
}
\description{
grab image in clipboard (if have), and save to a local filepath
}
