% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{help_proxy}
\alias{help_proxy}
\title{Start a passthrough proxy for R's help pages}
\usage{
help_proxy(port = NULL, host = "0.0.0.0")
}
\arguments{
\item{port}{Candidate proxy ports.}

\item{host}{Bind address for the proxy (\code{"127.0.0.1"} or \code{"0.0.0.0"}).}
}
\value{
The help index URL, invisibly.
}
\description{
Start a proxy that forwards all request paths to R's internal httpd, which
can be used to expose the help system to other devices on LAN.
}
