% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{is_R_CMD_check}
\alias{is_R_CMD_check}
\alias{is_CRAN_incoming}
\alias{check_package_name}
\alias{check_old_package}
\title{Some utility functions for checking packages}
\usage{
is_R_CMD_check()

is_CRAN_incoming()

check_package_name()

check_old_package(name, version)
}
\description{
Miscellaneous utility functions to obtain information about the package
checking environment.
}
\keyword{internal}
