% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{mark_dirs}
\alias{mark_dirs}
\title{Mark some paths as directories}
\usage{
mark_dirs(x)
}
\arguments{
\item{x}{Character vector of paths to files and directories.}
}
\description{
Add a trailing backlash to a file path if this is a directory. This is useful
in messages to the console for example to quickly identify directories from
files.
}
\details{
If \code{x} is a vector of relative paths, directory test is done with
path relative to the current working dir. Use \code{xfun::\link{in_dir}()} or
use absolute paths.
}
\examples{
mark_dirs(list.files(find.package("xfun"), full.names = TRUE))
}
