% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{md_table}
\alias{md_table}
\title{Generate a Markdown pipe table}
\usage{
md_table(x, digits = NULL, caption = NULL, na = "", newline = " ")
}
\arguments{
\item{x}{A 2-dimensional object (e.g., a matrix or data frame).}

\item{digits}{The number of decimal places to be passed to \code{\link[=round]{round()}}. It can
be a integer vector of the same length as the number of columns in \code{x} to
round columns separately.}

\item{caption}{A table caption. If provided, the caption will be placed in a
fenced Div with the class \code{tab-caption}, and the whole table will be placed
in a fenced Div with the class \code{table}.}

\item{na}{A character string to represent \code{NA} values.}

\item{newline}{A character string to substitute \verb{\\n} in \code{x} (because pipe
tables do not support line breaks in cells).}
}
\value{
A character vector.
}
\description{
A minimal Markdown table generator using the pipe \code{|} as column separators.
}
\examples{
xfun::md_table(head(iris))
xfun::md_table(head(iris), caption = "An old dataset.")
}
\seealso{
\code{\link[knitr:kable]{knitr::kable()}} (which supports more features)
}
