% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{optipng}
\alias{optipng}
\title{Run OptiPNG on all PNG files under a directory}
\usage{
optipng(
  dir = ".",
  files = list.files(dir, "[.]png$", recursive = TRUE, full.names = TRUE),
  ...
)
}
\arguments{
\item{dir}{Path to a directory.}

\item{files}{Alternatively, you can choose the specific files to optimize.}

\item{...}{Arguments to be passed to \code{system2()}.}
}
\description{
Call the command \command{optipng} via \code{system2()} to optimize all PNG
files under a directory.
}
\references{
OptiPNG: \url{http://optipng.sourceforge.net}.
}
