% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{is_windows}
\alias{is_windows}
\alias{is_unix}
\alias{is_macos}
\alias{is_linux}
\alias{is_arm64}
\title{Test for types of operating systems}
\usage{
is_windows()

is_unix()

is_macos()

is_linux()

is_arm64()
}
\description{
Functions based on \code{.Platform$OS.type} and \code{Sys.info()} to test if
the current operating system is Windows, macOS, Unix, or Linux.
}
\examples{
library(xfun)
# only one of the following statements should be true
is_windows()
is_unix() && is_macos()
is_linux()
# In newer Macs, CPU can be either Intel or Apple
is_arm64()  # TRUE on Apple silicone machines
}
