% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_only}
\alias{parse_only}
\title{Parse R code and do not keep the source}
\usage{
parse_only(code)
}
\arguments{
\item{code}{A character vector of the R source code.}
}
\value{
R \code{\link{expression}}s.
}
\description{
An abbreviation of \code{parse(keep.source = FALSE)}.
}
\examples{
library(xfun)
parse_only("1+1")
parse_only(c("y~x", "1:5 # a comment"))
parse_only(character(0))
}
