% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{post_release}
\alias{post_release}
\title{Perform post-CRAN-release routine tasks}
\usage{
post_release()
}
\description{
After a CRAN release is accepted, commit and tag the release, bump the
package version to the next patch dev version, prepend a new header to
\file{NEWS.md}, push to the remote repository, and create a GitHub release
with the news items of this version.
}
\details{
The repo will be tagged with the version number extracted from the \code{Version}
field in \file{DESCRIPTION}. Then \code{.1} is appended to \code{Version} in
\file{DESCRIPTION} as the next patch version, and a new header for the next
release version (which is from bumping the last digit of the current version)
is prepended to \file{NEWS.md}. For example, if the current version is
\verb{1.2.3} (or \code{1.2}), the next patch version will be \verb{1.2.3.1} (or \verb{1.2.1}) and
the next release version will be \verb{1.2.4} (or \code{1.3}).
}
