% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{proc_kill}
\alias{proc_kill}
\title{Kill a process and (optionally) all its child processes}
\usage{
proc_kill(pid, recursive = TRUE, ...)
}
\arguments{
\item{pid}{The process ID.}

\item{recursive}{Whether to kill the child processes of the process.}

\item{...}{Arguments to be passed to \code{\link[=system2]{system2()}} to run the
command to kill the process.}
}
\value{
The status code returned from \code{system2()}.
}
\description{
Run the command \command{taskkill /f /pid} on Windows and \command{kill} on
Unix, respectively, to kill a process.
}
