% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record_print}
\alias{record_print}
\alias{record_print.default}
\title{Print methods for \code{record()}}
\usage{
record_print(x, ...)

\method{record_print}{default}(x, ...)
}
\arguments{
\item{x}{The value to be printed.}

\item{...}{Other arguments to be passed to \code{record_print()} methods.}
}
\value{
A \code{record_print()} method should return a character vector. The
returned value may have a special class \code{record_asis}, which will be stored
in the \code{record()} output. All other classes will be discarded.
}
\description{
An S3 generic function to be called to print visible values in code when the
code is recorded by \code{\link[=record]{record()}}. It is similar to \code{\link[knitr:knit_print]{knitr::knit_print()}}. By
default, it captures the normal \code{\link[=print]{print()}} output and returns the result as a
character vector. Users and package authors can define other S3 methods to
extend this function.
}
