% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{Rscript}
\alias{Rscript}
\alias{Rcmd}
\title{Run the commands \command{Rscript} and \command{R CMD}}
\usage{
Rscript(args, ...)

Rcmd(args, ...)
}
\arguments{
\item{args}{A character vector of command-line arguments.}

\item{...}{Other arguments to be passed to \code{\link{system2}()}.}
}
\value{
A value returned by \code{system2()}.
}
\description{
Wrapper functions to run the commands \command{Rscript} and \command{R CMD}.
}
\examples{
library(xfun)
Rscript(c("-e", "1+1"))
Rcmd(c("build", "--help"))
}
