% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{same_path}
\alias{same_path}
\title{Test if two paths are the same after they are normalized}
\usage{
same_path(p1, p2, ...)
}
\arguments{
\item{p1, p2}{Two vectors of paths.}

\item{...}{Arguments to be passed to \code{\link{normalize_path}()}.}
}
\description{
Compare two paths after normalizing them with the same separator (\code{/}).
}
\examples{
library(xfun)
same_path("~/foo", file.path(Sys.getenv("HOME"), "foo"))
}
