% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stringsAsStrings}
\alias{stringsAsStrings}
\title{Set the global option \code{\link{options}(stringsAsFactors = FALSE)} inside
a parent function and restore the option after the parent function exits}
\usage{
stringsAsStrings()
}
\description{
This is a shorthand of \code{opts = options(stringsAsFactors = FALSE);
on.exit(options(opts), add = TRUE)}.
}
\keyword{internal}
