% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{strip_html}
\alias{strip_html}
\title{Strip HTML tags}
\usage{
strip_html(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with HTML tags and comments stripped off.
}
\description{
Remove HTML tags and comments from text.
}
\examples{
xfun::strip_html("<a href=\"#\">Hello <!-- comment -->world!</a>")
}
