% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{system3}
\alias{system3}
\title{A customized wrapper for \code{system2()}}
\usage{
system3(command, args = character(), ...)
}
\arguments{
\item{command, args, ...}{Passed to \code{\link[=system2]{system2()}}. Note that \code{args} will always
be quoted with \code{\link[=shQuote]{shQuote()}} before passing to \code{system2()}, so you don't need
to quote it yourself.}
}
\value{
The value returned by \code{system2()}.
}
\description{
If \code{system2()} returns character output (e.g., with the argument \code{stdout = TRUE}), check if the output is encoded in UTF-8. If it is, mark it with UTF-8
explicitly. If it returns an integer, treat it as an error code and throw an
error if it is non-zero.
}
\examples{\dontshow{if (interactive()) \{}
a = c("-e", "print(intToUtf8(c(20320, 22909)))")
x2 = system2("Rscript", shQuote(a), stdout = TRUE)
Encoding(x2)  # unknown

x3 = xfun::system3("Rscript", a, stdout = TRUE)
# encoding of x3 should be UTF-8 if the current locale is UTF-8
!l10n_info()[["UTF-8"]] || Encoding(x3) == "UTF-8"  # should be TRUE
\dontshow{\}}
}
