% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{taml_load}
\alias{taml_load}
\alias{taml_file}
\title{A simple YAML parser}
\usage{
taml_load(x, envir = parent.frame())

taml_file(x)
}
\arguments{
\item{x}{For \code{taml_load()}, a character vector of the YAML content. For
\code{taml_file()}, a file path.}

\item{envir}{The environment in which R expressions in YAML are evaluated. To
disable the evaluation, use \code{envir = FALSE}.}
}
\value{
A list.
}
\description{
TAML is a tiny subset of YAML. See
\url{https://yihui.org/litedown/#sec:yaml-syntax} for its specifications.
}
\examples{
taml_load("a: 1")
taml_load("a: 1\nb: \"foo\"\nc: null")
taml_load("a:\n  b: false\n  c: true\n  d: 1.234\ne: bar")
taml_load("a: !expr paste(1:10, collapse = \", \")")
taml_load("a: [1, 3, 4, 2]")
taml_load("a: [1, \"abc\", 4, 2]")
taml_load("a: [\"foo\", \"bar\"]")
taml_load("a: [true, false, true]")
# the other form of array is not supported
taml_load("a:\n  - b\n  - c")
# and you must use the yaml package
if (loadable("yaml")) yaml_load("a:\n  - b\n  - c")
}
