% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_silent}
\alias{try_silent}
\title{Try to evaluate an expression silently}
\usage{
try_silent(expr)
}
\arguments{
\item{expr}{An R expression.}
}
\description{
An abbreviation of \code{try(silent = TRUE)}.
}
\examples{
library(xfun)
z = try_silent(stop("Wrong!"))
inherits(z, "try-error")
}
