% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{url_accessible}
\alias{url_accessible}
\title{Test if a URL is accessible}
\usage{
url_accessible(x)
}
\arguments{
\item{x}{A URL as a character string.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Try to download a URL via \code{xfun::\link{download_file}()} and see if it
succeeds.
}
\examples{
xfun::url_accessible("https://yihui.org")
}
