% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_encode}
\alias{base64_encode}
\title{Encode data into the base64 encoding.}
\usage{
base64_encode(x)
}
\arguments{
\item{x}{A raw vector. If not raw, it is assumed to be a file or a connection
to be read as raw via \code{readBin()}.}
}
\value{
A character string.
}
\description{
Encode a file or a raw vector into the base64 encoding.
}
\examples{
xfun::base64_encode(as.raw(1:10))
logo = file.path(R.home("doc"), "html", "logo.jpg")
xfun::base64_encode(logo)
}
