% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_uri}
\alias{base64_uri}
\title{Generate the Data URI for a file}
\usage{
base64_uri(x)
}
\arguments{
\item{x}{A file path.}
}
\value{
A string of the form \verb{data:<media type>;base64,<data>}.
}
\description{
Encode the file in the base64 encoding, and add the media type. The data URI
can be used to embed data in HTML documents, e.g., in the \code{src}
attribute of the \verb{<img />} tag.
}
\examples{
logo = file.path(R.home("doc"), "html", "logo.jpg")
img = htmltools::img(src = xfun::base64_uri(logo), alt = "R logo")
if (interactive()) htmltools::browsable(img)
}
