% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{is_abs_path}
\alias{is_abs_path}
\alias{is_rel_path}
\title{Test if paths are relative or absolute}
\usage{
is_abs_path(x)

is_rel_path(x)
}
\arguments{
\item{x}{A vector of paths.}
}
\value{
A logical vector.
}
\description{
On Unix, check if the paths start with \file{/} or \file{~} (if they do, they
are absolute paths). On Windows, check if the paths start with a drive letter
(e.g., \file{C:}), a (back)slash (e.g., \file{/foo}), a UNC prefix (e.g.,
\file{\\\\host\\share}), or \file{~}.
}
\examples{
xfun::is_abs_path(c("C:/foo", "foo.txt", "/Users/john/", tempdir()))
xfun::is_rel_path(c("C:/foo", "foo.txt", "/Users/john/", tempdir()))
}
