% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isFALSE}
\alias{isFALSE}
\title{Test if an object is identical to \code{FALSE}}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
A simple abbreviation of \code{identical(x, FALSE)}.
}
\examples{
library(xfun)
isFALSE(TRUE)  # false
isFALSE(FALSE)  # true
isFALSE(c(FALSE, FALSE))  # false
}
