% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{normalize_path}
\alias{normalize_path}
\title{Normalize paths}
\usage{
normalize_path(x, winslash = "/", must_work = FALSE)
}
\arguments{
\item{x, winslash, must_work}{Arguments passed to
\code{\link{normalizePath}()}.}
}
\description{
A wrapper function of \code{normalizePath()} with different defaults.
}
\examples{
library(xfun)
normalize_path("~")
}
