% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{optipng}
\alias{optipng}
\title{Run OptiPNG on all PNG files under a directory}
\usage{
optipng(dir = ".")
}
\arguments{
\item{dir}{Path to a directory.}
}
\description{
Calls the command \command{optipng} to optimize all PNG files under a
directory.
}
\references{
OptiPNG: \url{http://optipng.sourceforge.net}.
}
