% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{post_release}
\alias{post_release}
\title{Perform post-CRAN-release routine tasks}
\usage{
post_release()
}
\description{
After a CRAN release is accepted, commit and tag the release, bump the
package version to the next patch dev version, prepend a new header to
\file{NEWS.md}, push to the remote repository, and create a GitHub release
with the news items of this version.
}
\details{
The version number \code{X.Y} or \code{X.Y.0} is extracted from the \code{Version} field in
\file{DESCRIPTION}. For example, if the current version is \code{0.58} (or
\verb{0.58.0}), the repo will be tagged as \code{v0.58} (or \code{v0.58.0}). The patch
version is then bumped to \verb{0.58.1}, and a new header for the next minor
version \code{0.59} (or \verb{0.59.0}) is prepended to \file{NEWS.md}.
}
