% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{post_release}
\alias{post_release}
\title{Perform post-CRAN-release routine tasks}
\usage{
post_release()
}
\description{
After a CRAN release is accepted, commit and tag the release, bump the
package version to the next patch dev version, prepend a new header to
\file{NEWS.md}, push to the remote repository, and create a GitHub release
with the news items of this version.
}
\details{
The version number \samp{X.Y} is extracted from the \samp{Version} field in
\file{DESCRIPTION}. For example, if the current version is \samp{0.58}, the
repo will be tagged as \samp{v0.58}. The patch version is then bumped to
\samp{0.58.1}, and a new header is prepended to \file{NEWS.md}.
}
