% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record_print}
\alias{record_print}
\alias{record_print.default}
\alias{record_print.knitr_kable}
\alias{record_new}
\title{Print methods for \code{record()}}
\usage{
record_print(x, ...)

\method{record_print}{default}(x, ...)

\method{record_print}{knitr_kable}(x, ...)

record_new(x, class)
}
\arguments{
\item{x}{For \code{record_print()}, the value to be printed. For \code{record_new()}, a
character vector to be included in the printed results.}

\item{...}{Other arguments to be passed to \code{record_print()} methods.}

\item{class}{A class name. Possible values are:
\Sexpr{paste(xfun:::.record_cls, collapse = ', ')}.}
}
\value{
A \code{record_print()} method should return a character vector. The
original classes of the vector will be discarded, and the vector will be
treated as console output. If it should be another type of output, wrap the
vector in \code{\link[=record_new]{record_new()}} and specify a class name.
}
\description{
An S3 generic function to be called to print visible values in code when the
code is recorded by \code{\link[=record]{record()}}. It is similar to \code{\link[knitr:knit_print]{knitr::knit_print()}}. By
default, it captures the normal \code{\link[=print]{print()}} output and returns the result as a
character vector. The \code{knitr_kable} method is for printing \code{\link[knitr:kable]{knitr::kable()}}
output. Users and package authors can define other S3 methods to extend this
function.
}
